backend/pyproject.toml
[project]
name = "fins"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "anthropic>=0.75.0",
    "asyncpg>=0.30.0",
    "duckdb>=1.4.3",
    "fastapi>=0.125.0",
    "litellm>=1.80.10",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "passlib>=1.7.4",
    "pdfplumber>=0.11.8",
    "prefect>=3.6.6",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pypdf2>=3.0.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "python-levenshtein>=0.27.3",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.21",
    "tortoise-orm>=0.21.7",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
